Remove text file detection because it already exists
Once upon a time, I wanted to set textwidth=79 for .txt files, so I
created a ftplugin/txt.vim file to do that. What slipped through is that
this type of file is called "text", not "txt", so I had to make an
autocmd that sets the *.txt filetype to text.

This is not needed, as I can just use the real name of the file, and
that is "text". For instance, filetype for *.js is "javascript" etc.